feat(api): implement user and API key updates with proper validation and services
