+++ packages/api/src/user/user.service.ts
@@ -1,37 +1,15 @@
 import { ENV } from '../config/env_variables';
 import { ConflictError, NotFoundError } from '../utils/errors.utils';
 import { generateUUID } from '../utils/uuid.utils';
-import { IUserModel } from './user.model';
+import { IUserModel } from './interfaces/iuser.model';
+import { IUserService } from './interfaces/iuser.service';
 import {
   User,
   UserResponsePublic,
   AuthenticationRegisterRequestPublicSchema,
+  UserUpdate,
 } from '@dev-dashboard/shared';
 import bcrypt from 'bcryptjs';
-
-export interface IUserService {
-  create(
-    user: AuthenticationRegisterRequestPublicSchema
-  ): Promise<UserResponsePublic>;
-  findById(userId: string): Promise<UserResponsePublic>;
-  findByEmailPrivate(email: string): Promise<User>;
-  findByEmailPublic(email: string): Promise<UserResponsePublic>;
-  emailExists(email: string): Promise<boolean>;
-  update(
-    userId: string,
-    updates: Partial<Omit<User, 'userId' | 'email' | 'createdAt'>>
-  ): Promise<UserResponsePublic>;
-  delete(userId: string): Promise<void>;
-  updateLastLogin(
-    userId: string,
-    timestamp: string
-  ): Promise<UserResponsePublic>;
-  updatePassword(
-    userId: string,
-    passwordHash: string
-  ): Promise<UserResponsePublic>;
-  deactivate(userId: string): Promise<UserResponsePublic>;
-}
 
 export const UserService = (userModel: IUserModel): IUserService => {
   return {
@@ -144,9 +122,7 @@ export const UserService = (userModel: IUserModel): IUserService => {
 
     async update(
       userId: string,
-      updates: Partial<
-        Omit<User, 'userId' | 'email' | 'createdAt' | 'passwordHash'>
-      >
+      updates: UserUpdate
     ): Promise<UserResponsePublic> {
       try {
         const result = await userModel.update(userId, updates);
